fix quantifiers being too greedy
The issue was the quantifiers always matched as much as possible regardless of the following patterns. This meant that patterns like `[^a]+, b` didn't work properly.
For example above should match "bb, b" but it didn't because `[^a]+` matched the whole input and following `, b` didn't have anything to match and thus failed the match.However + means one or more, so it's totally valid for it to only match "bb" and thus following literal `, b` can match the rest of input.

The solution was to mostly rewrite the matcher to use global (to a match) state and recursion.

Other changes include:
  * groups are not nested patterns anymore, so that they won't mess with quantifiers
     The issue with this is that repeated groups don't work anymore (eg patterns like `(a)+`)
  * unified quantifiers to a single pattern element